+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
-All notable changes to this project will be documented in this file.
+
+## [0.2.0] - 2026-02-06
+### Added
+- Virtual Keypad
+- Display and Keypad CSS files
+
+### Changed
+- Added global event listener at container level
+- Migrate logic from Display to Calculator to support the above
+- Refactored and unified physical (keyboard) with virtual (keypad) keys
 
 ## [0.1.0] - 2026-02-04
 ### Added